+++ src/Core.php
@@ -5,18 +5,71 @@ use Olive\Tools;
 use Olive\UDMS\Common as Common;
 use Olive\UDMS\Exception\Custom as UException;
 use Olive\UDMS\Model\Database as Database;
-class core
+class Core
 {
     use Common;
 
+    private $path;
+
+    private $udmsCacheDir;
+
     private $selectedAddon;
 
-    private $execute;
+    public $execute;
+
+    private $d2tMode = false;
+
+    public $prefix = '';
+
+    private $cacheAppModel;
+
+    private $cacheAppModelData;
+
+    private $cacheD2T;
+
+    public $debugc = 0;
 
     // quickly function
+
+    private function setPath($path = null)
+    {
+        if ($path != null) {
+            $e = str_split($path);
+            if ($e[count($e) - 1] != '/') {
+                $path = $path . '/';
+            }
+            $this->path = $path;
+        } else {
+            $this->path = realpath(dirname(__FILE__) . '/../') . '/';
+        }
+    }
+
+    public function getPath($value = '')
+    {
+        return $this->path . $value;
+    }
+
+    private function setUCPath($path = null)
+    {
+        if ($path != null) {
+            $e = str_split($path);
+            if ($e[count($e) - 1] != '/') {
+                $path = $path . '/';
+            }
+            $this->udmsCacheDir = $path;
+        } else {
+            $this->udmsCacheDir = $this->getPath();
+        }
+    }
+
+    public function getUCPath($value = '')
+    {
+        return $this->udmsCacheDir . $value;
+    }
+
     public static function validName($value = '')
     {
-        return preg_match('/^([a-zA-Z])+(_)*([0-9a-zA-Z])*$/m', $value);
+        return preg_match('/^[a-zA-Z]+[0-9a-zA-Z_]*$/m', $value);
     }
 
     private function createDatabaseDir($name)
@@ -52,12 +105,9 @@ class core
     {
         $type2 = '\\Olive\\UDMS\\Addon\\' . $type . '\\Point';
         if (! class_exists($type2)) {
-            throw new UException($this->getUCPath(), $type2 . ' is not found!');
+            throw new UException($this->getUCPath(), $type2 . ' is not found.', 101);
         }
-        if (isset($GLOBALS['__udms_global'])) {
-            unset($GLOBALS['__udms_global']);
-        }
-        $this->execute = new $type2($this->path, $this->udmsCacheDir, $option);
+        $this->execute = new $type2($this, $option);
         $this->selectedAddon = $type;
     }
 
@@ -66,7 +116,7 @@ class core
         return $this->selectedAddon;
     }
 
-    public function validAppDataModel($model = [])
+    public function validAppModel($model = [])
     {
         if (! is_array($model)) {
             return false;
@@ -125,37 +175,113 @@ class core
         return true;
     }
 
-    public function setAppDataModel($model = [])
+    public function getAppModel()
     {
-        if (! $this->validAppDataModel($model)) {
-            throw new UException($this->getUCPath(), 'your data model is not valid!');
+        return $this->cacheAppModel;
+    }
+
+    public function getAppModelData()
+    {
+        return $this->cacheAppModelData;
+    }
+
+    public function setAppModel($model = [])
+    {
+        if (! $this->validAppModel($model)) {
+            throw new UException($this->getUCPath(), 'your data model is not valid!', 102);
         }
-        foreach ($model as $db => $di) {
-            $this->createDatabaseDir($db);
-            $this->updateDatabaseModel($db, $di);
-            $dudf = $this->getUCPath($db . '/appModelData.json');
-            if (! file_exists($dudf)) {
-                $this->updateDatabaseModelData($db);
-            }
+        Tools::file($this->getUCPath('appModel.json'), Tools::jsonEncode($model));
+        Tools::file($this->getUCPath('appModelData.json'), Tools::jsonEncode([]));
+        $this->resetAppModel();
+    }
+
+    public function setAppModelData($model = [])
+    {
+        $this->cacheAppModelData = $model;
+        Tools::file($this->getUCPath('appModelData.json'), Tools::jsonEncode($model));
+    }
+
+    private function resetAppModel()
+    {
+        $this->cacheAppModel = Tools::getJsonFile($this->getUCPath('appModel.json'));
+        $this->cacheAppModelData = Tools::getJsonFile($this->getUCPath('appModelData.json'));
+    }
+
+    public function getDatabaseModel($name)
+    {
+        if (isset($this->cacheAppModel[$name])) {
+            return $this->cacheAppModel[$name];
         }
+
+        return [];
+    }
+
+    public function getDatabaseModelData($name)
+    {
+        if (isset($this->cacheAppModelData[$name])) {
+            return $this->cacheAppModelData[$name];
+        }
+
+        return [];
+    }
+
+    public function updateDatabaseModel($name, $data = [])
+    {
+        $this->cacheAppModel[$name] = $data;
+        Tools::file($this->getUCPath('appModel.json'), Tools::jsonEncode($this->cacheAppModel));
+    }
+
+    public function updateDatabaseModelData($name, $data = [])
+    {
+        $this->cacheAppModelData[$name] = $data;
+        Tools::file($this->getUCPath('appModelData.json'), Tools::jsonEncode($this->cacheAppModelData));
+    }
+
+    public function addLog($dec = '', $file = '', $line = 0)
+    {
+        Tools::file($this->getUCPath('error_logs'), '[' . date('c', time()) . "][$file:$line]: $dec\n", 'a+');
+        echo $dec . "\n";
     }
 
     public function render()
     {
         if ($this->getAddon() == '') {
-            throw new UException($this->getUCPath(), 'render only with set addon. please first set your selection addon and next render!');
+            throw new UException($this->getUCPath(), 'render only with set addon. please first set your selection addon and next render!', 103);
         }
-        $dbs = Tools::getDirList($this->getUCPath());
-        foreach ($dbs as $db) {
-            $ui = $this->getDatabaseModel($db);
-            if (! $this->validAppDataModel([$db => $ui])) {
-                throw new UException($this->getUCPath(), 'your data model is not valid!');
+        $dbs = $this->getAppModel();
+
+        // first check AppModel
+        foreach ($dbs as $db => $ui) {
+            if (! $this->validAppModel([$db => $ui])) {
+                throw new UException($this->getUCPath(), 'your data model is not valid!', 104);
             }
             if (! isset($ui['__udms_config'])) {
                 $ui['__udms_config'] = [];
             }
             if (! $this->validName($db)) {
-                throw new UException($this->getUCPath(), $db . 'name is not valid!');
+                throw new UException($this->getUCPath(), $db . 'name is not valid!', 105);
+            }
+            unset($ui['__udms_config']);
+            foreach ($ui as $table => $ti) {
+                if (! isset($ti['__udms_config'])) {
+                    $ti['__udms_config'] = [];
+                }
+                if (! $this->validName($table)) {
+                    throw new UException($this->getUCPath(), $table . 'name is not valid!', 106);
+                }
+                unset($ti['__udms_config']);
+                foreach ($ti as $col => $ci) {
+                    if (! $this->validName($col)) {
+                        throw new UException($this->getUCPath(), $col . 'name is not valid!', 107);
+                    }
+                }
+            }
+        }
+
+        // final render
+        foreach ($dbs as $db => $ui) {
+            if (! isset($ui['__udms_config'])) {
+                $ui['__udms_config'] = [];
             }
             $this->createDatabase($db, $ui['__udms_config']);
             unset($ui['__udms_config']);
@@ -163,19 +289,56 @@ class core
                 if (! isset($ti['__udms_config'])) {
                     $ti['__udms_config'] = [];
                 }
-                if (! $this->validName($table)) {
-                    throw new UException($this->getUCPath(), $table . 'name is not valid!');
-                }
                 $this->$db->createTable($table, $ti['__udms_config']);
                 unset($ti['__udms_config']);
                 foreach ($ti as $col => $ci) {
-                    if (! $this->validName($col)) {
-                        throw new UException($this->getUCPath(), $col . 'name is not valid!');
-                    }
                     $this->$db->$table->createColumn($col, $ci);
                 }
             }
         }
+    }
+
+    public function setD2TMode($database = '')
+    {
+        if (! $this->validName($database)) {
+            throw new UException($this->getUCPath(), $database . ' name is not valid!', 108);
+        }
+        if ($this->inReservedName($database)) {
+            throw new UException($this->getUCPath(), 'your database name is reserved! (' . $database . ')', 109);
+        }
+        if (! $this->existsDatabase($database)) {
+            throw new UException($this->getUCPath(), 'database D2TMode name (' . $database . ') not exists.', 110);
+        }
+
+        $this->createDatabaseDir($database);
+        $this->d2tDatabase = $database;
+        $this->d2t = new Database($this, $database);
+        $this->d2tMode = true;
+    }
+
+    public function desD2TMode()
+    {
+        $this->d2tDatabase = '';
+        unset($this->d2t);
+        $this->d2tMode = false;
+        $this->prefix = '';
+        $this->resetAppModel();
+        $this->setD2T();
+    }
+
+    public function getD2T()
+    {
+        if (is_null($this->cacheD2T)) {
+            $this->cacheD2T = Tools::getJsonFile($this->getUCPath('d2t.json'));
+        }
+
+        return $this->cacheD2T;
+    }
+
+    private function setD2T($data = [])
+    {
+        $this->cacheD2T = $data;
+        Tools::file(($this->getUCPath('d2t.json')), Tools::jsonEncode($data));
     }
 
     public function availableDatabaseRule()
@@ -192,7 +355,11 @@ class core
 
     public function listDatabases()
     {
-        $dl = $this->execute->listDatabases();
+        if ($this->d2tMode == false) {
+            $dl = $this->execute->listDatabases();
+        } else {
+            $dl = $this->getD2T();
+        }
         if (is_array($dl)) {
             return $dl;
         } else {
@@ -202,84 +369,92 @@ class core
 
     public function createDatabase($name, $option = [])
     {
-        if (! $this->validName($name)) {
-            throw new UException($this->getUCPath(), $name . 'name is not valid!');
+        if ($this->existsDatabase($name)) {
+            throw new UException($this->getUCPath(), 'your database name has exists (' . $name . ')!', 111);
         }
-        if (! $this->existsDatabase($name)) {
+        if ($this->d2tMode == false) {
             $this->createDatabaseDir($name);
             $this->execute->createDatabase($name, $option);
-            $didf = $this->getUCPath($name . '/appModel.json');
-            if (! file_exists($didf)) {
-                if (count($option) > 0) {
-                    $dm = [
-                    '__udms_config' => $option
-                  ];
-                } else {
-                    $dm = [];
-                }
-                $this->updateDatabaseModel($name, $dm);
-            }
-            $dudf = $this->getUCPath($name . '/appModelData.json');
-            if (! file_exists($dudf)) {
-                $this->updateDatabaseModelData($name);
-            }
         } else {
-            throw new UException($this->getUCPath(), 'your database name has exists (' . $name . ')!');
+            $dl = $this->getD2T();
+            $dl[] = $name;
+            $this->setD2T($dl);
         }
     }
 
-    public function dropDatabase($name, $keep = false)
+    public function dropDatabase($name)
     {
-        if (! $this->validName($name)) {
-            throw new UException($this->getUCPath(), $name . 'name is not valid!');
-        }
         if ($this->existsDatabase($name)) {
-            $this->execute->dropDatabase($name);
-            if ($keep == false) {
+            if ($this->d2tMode == false) {
+                $this->execute->dropDatabase($name);
                 $this->dropDatabaceDir($name);
+            } else {
+                $dtl = $this->d2t->listTables();
+                foreach ($dtl as $dt) {
+                    $this->d2t->dropTable($dt);
+                }
+                $this->setD2T(array_diff($this->getD2T(), [$name]));
             }
         } else {
-            throw new UException($this->getUCPath(), 'your database name has not exists (' . $name . ')');
+            throw new UException($this->getUCPath(), 'your database name has not exists (' . $name . ')', 112);
         }
     }
 
     public function existsDatabase($name)
     {
         if (! $this->validName($name)) {
-            throw new UException($this->getUCPath(), $name . 'name is not valid!');
+            throw new UException($this->getUCPath(), $name . 'name is not valid!', 113);
         }
-
-        return $this->execute->existsDatabase($name);
+        if ($this->d2tMode == false) {
+            return $this->execute->existsDatabase($name);
+        } else {
+            return in_array($name, $this->listDatabases());
+        }
     }
 
     public function renameDatabase($name, $to)
     {
-        if (! $this->validName($name)) {
-            throw new UException($this->getUCPath(), $name . 'name is not valid!');
-        }
-        if (! $this->validName($to)) {
-            throw new UException($this->getUCPath(), $to . 'name is not valid!');
-        }
         if ($this->existsDatabase($to)) {
-            throw new UException($this->getUCPath(), 'your database name exists (' . $to . ')');
+            throw new UException($this->getUCPath(), 'your database name has exists (' . $to . ')', 115);
         }
         if (! $this->existsDatabase($name)) {
-            throw new UException($this->getUCPath(), 'your database name has not exists (' . $name . ')');
+            throw new UException($this->getUCPath(), 'your database name has not exists (' . $name . ')', 114);
         }
-        $this->execute->renameDatabase($name, $to);
-        $this->renameDatabaceDir($name, $to);
+        if ($this->d2tMode == false) {
+            $this->execute->renameDatabase($name, $to);
+            $this->renameDatabaceDir($name, $to);
+        } else {
+            $dtl = $this->d2t->listTables();
+            foreach ($dtl as $dt) {
+                if ($dt == $name) {
+                    $this->d2t->renameTable($name, $to);
+
+                    break;
+                }
+            }
+            $dl = $this->getD2T();
+            $key = array_search($name, $dl);
+            $dl[$key] = $to;
+            $this->setD2T($dl);
+        }
     }
 
     public function __get($name)
     {
         if ($this->inReservedName($name)) {
-            throw new UException($this->getUCPath(), 'your database name is reserved! (' . $name . ')');
+            throw new UException($this->getUCPath(), 'your database name is reserved! (' . $name . ')', 116);
         }
         if (! $this->existsDatabase($name)) {
-            throw new UException($this->getUCPath(), 'your database name can not found! (' . $name . ')');
+            throw new UException($this->getUCPath(), 'can not found your database name! (' . $name . ')', 117);
         }
+        $this->od = $name;
+        if ($this->d2tMode == false) {
+            return new Database($this, $name);
+        } else {
+            $this->prefix = $name . '_';
 
-        return new Database($name, $this->path, $this->udmsCacheDir, $this->execute);
+            return $this->d2t;
+        }
     }
 
     public function __construct($udmsCacheDir = null)
@@ -287,7 +462,7 @@ class core
         if ($udmsCacheDir != null) {
             $ec = dirname($udmsCacheDir);
             if (! is_dir($ec)) {
-                throw new UException($udmsCacheDir, 'Can not access your UMDS Cache directory path!');
+                throw new UException($udmsCacheDir, 'Can not access your UMDS Cache directory path!', 100);
             }
         }
         $this->setPath();
@@ -295,6 +470,15 @@ class core
         if (! is_dir($this->getUCPath())) {
             mkdir($this->getUCPath());
         }
+        $af = $this->getUCPath('appModel.json');
+        $adf = $this->getUCPath('appModelData.json');
+        if (! file_exists($af)) {
+            Tools::file($af, '[]');
+        }
+        if (! file_exists($adf)) {
+            Tools::file($adf, '[]');
+        }
+        $this->resetAppModel();
         $this->reservedName = get_class_methods($this);
     }
 }